+++ CHANGES.md
@@ -2,6 +2,8 @@
 
 ## Next
 
+* PR #668: Using spark default parallelism as default.
+
 ## 0.25.0 - 2022-05-31
 * Issue #593: Allow users to disable cache when loading data via SQL query,
   by setting `cacheExpirationTimeInMinutes=0`